Move contribution field to its place
